Varios cambios para mejorar el modelo, los adapters, los repos y mas. Se agregaron los demas adapters, pero hay un bug en el delete y en el update

+++ domain/src/main/java/co/edu/javeriana/domain/logic/EstudioLogic.java
@@ -11,5 +11,5 @@ public interface EstudioLogic {
 
     List<Estudio> findAll(int persistence);
     Estudio findByIdCc(Integer cc, Integer id,int persistence);
-    Integer count(int persistence);
+    Long count(int persistence);
 }
